requirements.txt
azure-ai-textanalytics==5.3.0
azure-common==1.1.28
azure-core==1.32.0
certifi==2024.12.14
charset-normalizer==3.4.1
idna==3.10
isodate==0.7.2
requests==2.32.3
six==1.17.0
typing_extensions==4.12.2
urllib3==2.3.0
